blogilistan laajennus, step9
T4.21 valmis

+++ osa4/blogilista/controllers/blogs.js
@@ -3,11 +3,23 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
+//kaikkien blogien hakeminen
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 })
 
+//yksittäisen blogin hakeminen
+blogsRouter.get('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog) {
+    response.json(blog.toJSON())
+  } else {
+    response.status(404).end()
+  }
+})
+
+//uuden blogin tekeminen
 blogsRouter.post('/', async (request, response) => {
   const body = new Blog(request.body)
   const token = request.token
@@ -36,20 +48,25 @@ blogsRouter.post('/', async (request, response) => {
   }
 })
 
-blogsRouter.get('/:id', async (request, response) => {
+//blogin poistaminen
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = await Blog.findById(request.params.id)
-  if (blog) {
-    response.json(blog.toJSON())
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
   } else {
-    response.status(404).end()
+    return response.status(401).json({ error: 'unauthorized action' })
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-})
-
+//blogin tykkäysten muokkaaminen
 blogsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
   if (body.title === undefined || body.url === undefined) {
